+++ Kernel/Modules/AgentTicketZoom.pm
@@ -967,7 +967,8 @@ sub MaskAgentZoom {
             # ('AD1', 'AD3', 'AD2')
 
             my @TmpActivityDialogList
-                = map { $NextActivityDialogs->{$_} } sort keys %{$NextActivityDialogs};
+                = map { $NextActivityDialogs->{$_} }
+                sort  { $a <=> $b } keys %{$NextActivityDialogs};
 
             # we have to check if the current user has the needed permissions to view the
             # different activity dialogs, so we loop over every activity dialog and check if there
@@ -1030,7 +1031,7 @@ sub MaskAgentZoom {
             );
 
             if ( IsHashRefWithData($NextActivityDialogs) ) {
-                for my $NextActivityDialogKey ( sort keys %{$NextActivityDialogs} ) {
+                for my $NextActivityDialogKey ( sort { $a <=> $b } keys %{$NextActivityDialogs} ) {
                     my $ActivityDialogData = $Self->{ActivityDialogObject}->ActivityDialogGet(
                         Interface              => 'AgentInterface',
                         ActivityDialogEntityID => $NextActivityDialogs->{$NextActivityDialogKey},